% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINALIZEaggr.R
\name{FINALIZEaggr}
\alias{FINALIZEaggr}
\title{Finalize sequence statitistics}
\usage{
FINALIZEaggr(v)
}
\arguments{
\item{aggr}{A matrix containing the existing, final aggregate statistics. Should be output of \code{UPDATEaggr}: \code{n}, \code{mean}, \code{M2}.}
}
\value{
A matrix with 1 row and 3 columns: Current size \code{n}, current mean \code{mean} and current unbiased variance, \code{var}.
}
\description{
This function converts \code{M2} to unbiased Variance \code{var} and exports the final statistics.
}
\examples{
seq <- matrix(sample(1:100, 100), ncol = 5)
m <- INIaggr(seq[1,])
for (i in 2:20){
m <- UPDATEaggr(m, seq[i,])
}

mfin <- FINALIZEaggr(m)
all(
all.equal(mfin[1,1], nrow(seq)),
all.equal(mfin[,2], colMeans(seq)),
all.equal(mfin[,3], apply(seq, 2, FUN = "var"))
)

}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
