% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INIaggr.R
\name{INIaggr}
\alias{INIaggr}
\title{Initiate a sequence of values to be updated online.}
\usage{
INIaggr(vec)
}
\arguments{
\item{vec}{A numeric value. This is the first value of a sequence.}
}
\value{
A matrix with 1 row and 3 columns: Sample size \code{n = 1}, current mean \code{mean = vec} and current M2, \code{M2 = 0}, which is the average distance to \code{vec} to \code{mean}, so initially 0.
}
\description{
This function initiates the output matrix for Welford's Online algorithm.
}
\examples{
seq <- c(1:10)
m <- INIrun(seq[1])
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
