% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPDATEaggr.R
\name{UPDATEaggr}
\alias{UPDATEaggr}
\title{Update sequence statitistics}
\usage{
UPDATEaggr(aggr, newvalue)
}
\arguments{
\item{aggr}{A matrix containing the existing aggregate statistics. Usually output of \code{INIaggr}: \code{n}, \code{mean}, \code{M2}.}
}
\value{
A matrix with 1 row and 3 columns: Current size \code{n}, current mean \code{mean} and current M2, \code{M2}, which is the average distance previous values to \code{mean}.
}
\description{
This function updates n, mean and M2 when a new value is added to the sequence.
}
\examples{
#' seq <- c(1:10)
m <- INIrun(seq[1])
for (i in 2:10){
m <- updatestats(m, seq[i])
}
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
