% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INIaggr.R
\name{INIaggr}
\alias{INIaggr}
\title{Initiate a sequence of values to be updated online.}
\usage{
INIaggr(vec)
}
\arguments{
\item{vec}{A numeric value. This is the first value of a sequence.}
}
\value{
A matrix with \code{length(vec)} rows and 3 columns: Sample size \code{n = 1}, current mean \code{mean[i] = vec[i]} and current M2, \code{M2 = 0}, which is the average distance to \code{vec[i]} to \code{mean[i]}, so initially 0.
}
\description{
This function initiates the output matrix for Welford's Online algorithm.
}
\examples{
seq <- matrix(sample(1:100, 100), ncol = 5)
m <- INIaggr(seq[1,])
}
\references{
Welford, B P. “Note on a Method for Calculating Corrected Sums of Squares and Products.” Technometrics 4, no. 3 (1962): 419–20.
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
