% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPDATEaggr.R
\name{UPDATEaggr}
\alias{UPDATEaggr}
\title{Update sequence statitistics}
\usage{
UPDATEaggr(aggr, newvalue)
}
\arguments{
\item{aggr}{A matrix containing the existing aggregate statistics. Usually output of \code{INIaggr}: \code{n}, \code{mean}, \code{M2}.}
}
\value{
A matrix with \code{nrow(aggr)} rows and 3 columns: Current size \code{n}, current mean \code{mean} and current M2, \code{M2}, which is the average distance previous values to \code{mean}.
}
\description{
This function updates n, mean and M2 when a new value is added to the sequence.
}
\examples{
seq <- matrix(sample(1:100, 100), ncol = 5)
m <- INIaggr(seq[1,])
for (i in 2:20){
m <- UPDATEaggr(m, seq[i,])
}
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
